requirements.txt
requests
PyYaml
python-telegram-bot
